move `FilterFloatButton` under /collection/activity folder
